NeoAI-codex-build-neurobridge-full-stack-web-app/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
openai==1.30.1
python-dotenv==1.0.1
requests==2.32.2
